// src/app/financeiro/contaspagar/enums/financeiro-contaspagar-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 3.4.1
Code generated at time stamp: 2019-05-30T20:20:55.617
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface FormaPagamento extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoContaBancaria extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}